+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 #Unreleased
-- [...]
+- NEW ICONS: DoubleArrow, Youtube, Instagram, Twitter, Trash
 
 #v0.1.0
 - [Transitions] Add Transitions component