+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Unreleased
-- NEW FakeTextField component
+- NEW UneditableTextField component
 - NEW ICONS: DoubleArrow, Youtube, Instagram, Twitter, Trash
 
 #v0.1.0