+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# Unreleased
+#v0.1.1
 - NEW UneditableTextField component
 - NEW ICONS: DoubleArrow, Youtube, Instagram, Twitter, Trash
+- [Modal] Fix render component on server side
 
 #v0.1.0
 - [Transitions] Add Transitions component